Fix autoload behaviour
Autoloading the defadvice is not advisable, since it takes effect even if a user chooses not to enable `wisp-mode`.

And autoloading a `defcustom` with the filename extension, then also modifying `auto-mode-alist` in an autoloaded block will not have the desired effect: it is unlikely that custom variables will have been set before the autoloads file has been executed, and hence any user customisation will have no effect. The standard solution - where the file extension is reasonably unambiguous - is to just unconditionally add it to `auto-mode-alist`.